+++ git-refspec/src/spec.rs
@@ -1,6 +1,5 @@
 use crate::types::Push;
 use crate::{Fetch, Instruction, Mode, Operation, RefSpec, RefSpecRef};
-use bstr::BStr;
 
 /// Access
 impl RefSpecRef<'_> {
@@ -11,28 +10,15 @@ impl RefSpecRef<'_> {
 
     /// Transform the state of the refspec into an instruction making clear what to do with it.
     pub fn instruction(&self) -> Instruction<'_> {
-        fn has_pattern(item: &BStr) -> bool {
-            item.contains(&b'*')
-        }
         match self.op {
             Operation::Fetch => match (self.mode, self.src, self.dst) {
                 (Mode::Normal | Mode::Force, Some(src), None) => Instruction::Fetch(Fetch::Only { src }),
-                (Mode::Normal | Mode::Force, Some(src), Some(dst)) if has_pattern(src) => {
-                    Instruction::Fetch(Fetch::AndUpdateMultipleWithGlob {
-                        src,
-                        dst,
-                        allow_non_fast_forward: matches!(self.mode, Mode::Force),
-                    })
-                }
-                (Mode::Normal | Mode::Force, Some(src), Some(dst)) => Instruction::Fetch(Fetch::AndUpdateSingle {
+                (Mode::Normal | Mode::Force, Some(src), Some(dst)) => Instruction::Fetch(Fetch::AndUpdate {
                     src,
                     dst,
                     allow_non_fast_forward: matches!(self.mode, Mode::Force),
                 }),
-                (Mode::Negative, Some(src), None) if has_pattern(src) => {
-                    Instruction::Fetch(Fetch::ExcludeMultipleWithGlob { src })
-                }
-                (Mode::Negative, Some(src), None) => Instruction::Fetch(Fetch::ExcludeSingle { src }),
+                (Mode::Negative, Some(src), None) => Instruction::Fetch(Fetch::Exclude { src }),
                 (mode, src, dest) => {
                     unreachable!(
                         "BUG: fetch instructions with {:?} {:?} {:?} are not possible",
@@ -41,7 +27,7 @@ impl RefSpecRef<'_> {
                 }
             },
             Operation::Push => match (self.mode, self.src, self.dst) {
-                (Mode::Normal | Mode::Force, Some(src), None) => Instruction::Push(Push::Single {
+                (Mode::Normal | Mode::Force, Some(src), None) => Instruction::Push(Push::Matching {
                     src,
                     dst: src,
                     allow_non_fast_forward: matches!(self.mode, Mode::Force),
@@ -52,22 +38,12 @@ impl RefSpecRef<'_> {
                 (Mode::Normal | Mode::Force, None, None) => Instruction::Push(Push::AllMatchingBranches {
                     allow_non_fast_forward: matches!(self.mode, Mode::Force),
                 }),
-                (Mode::Normal | Mode::Force, Some(src), Some(dst)) if has_pattern(src) => {
-                    Instruction::Push(Push::MultipleWithGlob {
-                        src,
-                        dst,
-                        allow_non_fast_forward: matches!(self.mode, Mode::Force),
-                    })
-                }
-                (Mode::Normal | Mode::Force, Some(src), Some(dst)) => Instruction::Push(Push::Single {
+                (Mode::Normal | Mode::Force, Some(src), Some(dst)) => Instruction::Push(Push::Matching {
                     src,
                     dst,
                     allow_non_fast_forward: matches!(self.mode, Mode::Force),
                 }),
-                (Mode::Negative, Some(src), None) if has_pattern(src) => {
-                    Instruction::Push(Push::ExcludeMultipleWithGlob { src })
-                }
-                (Mode::Negative, Some(src), None) => Instruction::Push(Push::ExcludeSingle { src }),
+                (Mode::Negative, Some(src), None) => Instruction::Push(Push::Exclude { src }),
                 (mode, src, dest) => {
                     unreachable!(
                         "BUG: push instructions with {:?} {:?} {:?} are not possible",
