+++ git-refspec/tests/parse/mod.rs
@@ -108,189 +108,8 @@ mod invalid {
     }
 }
 
-mod fetch {
-    use crate::parse::{assert_parse, b};
-    use git_refspec::{Fetch, Instruction, Mode};
-
-    #[test]
-    fn exclude_single() {
-        assert_parse("^a", Instruction::Fetch(Fetch::ExcludeSingle { src: b("a") }));
-    }
-
-    #[test]
-    fn exclude_multiple() {
-        assert_parse(
-            "^a*",
-            Instruction::Fetch(Fetch::ExcludeMultipleWithGlob { src: b("a*") }),
-        );
-    }
-
-    #[test]
-    fn lhs_colon_empty_fetches_only() {
-        assert_parse("src:", Instruction::Fetch(Fetch::Only { src: b("src") }));
-        let spec = assert_parse("+src:", Instruction::Fetch(Fetch::Only { src: b("src") }));
-        assert_eq!(
-            spec.mode(),
-            Mode::Force,
-            "force is set, even though it has no effect in the actual instruction"
-        );
-    }
-
-    #[test]
-    fn lhs_colon_rhs_updates_single_ref() {
-        assert_parse(
-            "a:b",
-            Instruction::Fetch(Fetch::AndUpdateSingle {
-                src: b("a"),
-                dst: b("b"),
-                allow_non_fast_forward: false,
-            }),
-        );
-        assert_parse(
-            "+a:b",
-            Instruction::Fetch(Fetch::AndUpdateSingle {
-                src: b("a"),
-                dst: b("b"),
-                allow_non_fast_forward: true,
-            }),
-        );
-    }
-
-    #[test]
-    fn lhs_colon_rhs_with_glob_updates_multiple_refs() {
-        assert_parse(
-            "a/*:b/*",
-            Instruction::Fetch(Fetch::AndUpdateMultipleWithGlob {
-                src: b("a/*"),
-                dst: b("b/*"),
-                allow_non_fast_forward: false,
-            }),
-        );
-        assert_parse(
-            "+a/*:b/*",
-            Instruction::Fetch(Fetch::AndUpdateMultipleWithGlob {
-                src: b("a/*"),
-                dst: b("b/*"),
-                allow_non_fast_forward: true,
-            }),
-        );
-    }
-
-    #[test]
-    fn empty_lhs_colon_rhs_fetches_head_to_destination() {
-        assert_parse(
-            ":a",
-            Instruction::Fetch(Fetch::AndUpdateSingle {
-                src: b("HEAD"),
-                dst: b("a"),
-                allow_non_fast_forward: false,
-            }),
-        );
-
-        assert_parse(
-            "+:a",
-            Instruction::Fetch(Fetch::AndUpdateSingle {
-                src: b("HEAD"),
-                dst: b("a"),
-                allow_non_fast_forward: true,
-            }),
-        );
-    }
-
-    #[test]
-    fn colon_alone_is_for_fetching_head_into_fetchhead() {
-        assert_parse(":", Instruction::Fetch(Fetch::Only { src: b("HEAD") }));
-        let spec = assert_parse("+:", Instruction::Fetch(Fetch::Only { src: b("HEAD") }));
-        assert_eq!(spec.mode(), Mode::Force, "it's set even though it's not useful");
-    }
-
-    #[test]
-    fn empty_refspec_is_enough_for_fetching_head_into_fetchhead() {
-        assert_parse("", Instruction::Fetch(Fetch::Only { src: b("HEAD") }));
-    }
-}
-
-mod push {
-    use crate::parse::{assert_parse, b};
-    use git_refspec::{Instruction, Mode, Push};
-
-    #[test]
-    fn exclude_single() {
-        assert_parse("^a", Instruction::Push(Push::ExcludeSingle { src: b("a") }));
-    }
-
-    #[test]
-    fn exclude_multiple() {
-        assert_parse("^a*", Instruction::Push(Push::ExcludeMultipleWithGlob { src: b("a*") }));
-    }
-
-    #[test]
-    fn lhs_colon_rhs_pushes_single_ref() {
-        assert_parse(
-            "a:b",
-            Instruction::Push(Push::Single {
-                src: b("a"),
-                dst: b("b"),
-                allow_non_fast_forward: false,
-            }),
-        );
-        assert_parse(
-            "+a:b",
-            Instruction::Push(Push::Single {
-                src: b("a"),
-                dst: b("b"),
-                allow_non_fast_forward: true,
-            }),
-        );
-    }
-
-    #[test]
-    fn lhs_colon_rhs_with_glob_pushes_multiple_refs() {
-        assert_parse(
-            "a/*:b/*",
-            Instruction::Push(Push::MultipleWithGlob {
-                src: b("a/*"),
-                dst: b("b/*"),
-                allow_non_fast_forward: false,
-            }),
-        );
-        assert_parse(
-            "+a/*:b/*",
-            Instruction::Push(Push::MultipleWithGlob {
-                src: b("a/*"),
-                dst: b("b/*"),
-                allow_non_fast_forward: true,
-            }),
-        );
-    }
-
-    #[test]
-    fn colon_alone_is_for_pushing_matching_refs() {
-        assert_parse(
-            ":",
-            Instruction::Push(Push::AllMatchingBranches {
-                allow_non_fast_forward: false,
-            }),
-        );
-        assert_parse(
-            "+:",
-            Instruction::Push(Push::AllMatchingBranches {
-                allow_non_fast_forward: true,
-            }),
-        );
-    }
-
-    #[test]
-    fn delete() {
-        assert_parse(":a", Instruction::Push(Push::Delete { ref_or_pattern: b("a") }));
-        let spec = assert_parse("+:a", Instruction::Push(Push::Delete { ref_or_pattern: b("a") }));
-        assert_eq!(
-            spec.mode(),
-            Mode::Force,
-            "force is set, even though it has no effect in the actual instruction"
-        );
-    }
-}
+mod fetch;
+mod push;
 
 mod util {
     use git_refspec::{Instruction, Operation, RefSpecRef};
